requirements.txt
django==1.8.4
Pillow
selenium
django-crispy-forms